requirements.txt
certifi==2025.10.5
charset-normalizer==3.4.4
contourpy==1.3.3
cycler==0.12.1
filelock==3.20.0
fonttools==4.60.1
fsspec==2025.9.0
idna==3.11
Jinja2==3.1.6
kiwisolver==1.4.9
MarkupSafe==3.0.2
matplotlib==3.10.7
mpmath==1.3.0
networkx==3.5
numpy==2.2.6
opencv-python==4.12.0.88
packaging==25.0
pillow==12.0.0
polars==1.34.0
polars-runtime-32==1.34.0
psutil==7.1.2
psycopg2-binary==2.9.11
pyparsing==3.2.5
python-dateutil==2.9.0.post0
PyYAML==6.0.3
redis==7.0.1
requests==2.32.5
scipy==1.16.2
six==1.17.0
sympy==1.14.0
torch==2.10.0.dev20251027
torchvision==0.25.0.dev20251027
typing_extensions==4.15.0
ultralytics==8.3.221
ultralytics-thop==2.0.17
urllib3==2.5.0
mediapipe>=0.10.0
insightface>=0.7.3
onnxruntime>=1.16.0
